… Backend/openapi/src/TmdbApi/lib/Model/TvSeriesDetails200Response.php