### Summary
- The media detail endpoint are now much faster
- Example:
- One Piece (+1000 episodes) from ~8 seconds to ~2 seconds
- Gute Zeiten Schlechte Zeiten (+8000 episodes) from ~9 seconds to ~ 2.5 seconds
- Added missing TMDB external id response model